+++ R/reexport-citation.R
@@ -1,0 +1,4 @@
+
+#' @importFrom utils citation
+#' @export
+utils::citation